src/modes/bookmark/requirements.txt
tables==0.0.1
plug[qplug]==0.0.1
